refactor(rollups-subgraph): Rename version folders from 0.9 with 1.0.
